feat: ready to tests

+++ controllers/cards.js
@@ -1,13 +1,13 @@
 const Card = require('../models/card');
-const { errorCodes } = require('../utils/errorCodes');
+const errorsList = require('../errors/index');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch((err) => res.status(errorCodes.internalServerError).send({ message: err.message }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const owner = req.user._id;
   const { name, link } = req.body;
 
@@ -15,32 +15,35 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(errorCodes.badRequest).send({ message: 'Переданы некорректные данные при создании карточки.' });
-      } else {
-        res.status(errorCodes.internalServerError).send({ message: err.name });
+        throw new errorsList.BadRequestError('Переданы некорректные данные при создании карточки.');
       }
-    });
+    })
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (card === null) {
-        res.status(errorCodes.notFound).send({ message: 'Карточка не найдена.' });
+        throw new errorsList.NotFoundError('Карточка не найдена.');
+      } else if (card.owner.toString() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => {
+            res.send({ data: card });
+          });
       } else {
-        res.send({ data: card });
+        throw new errorsList.UnauthorizedError('Это не ваша карточка!');
       }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorCodes.badRequest).send({ message: 'Переданы некорректные данные при создании карточки.' });
-      } else {
-        res.status(errorCodes.internalServerError).send({ message: err.message });
+        throw new errorsList.BadRequestError('Переданы некорректные данные при удалении карточки.');
       }
-    });
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -48,21 +51,20 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (card === null) {
-        res.status(errorCodes.notFound).send({ message: 'Карточка не найдена.' });
+        throw new errorsList.NotFoundError('Карточка не найдена.');
       } else {
         res.send({ data: card });
       }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorCodes.badRequest).send({ message: 'Переданы некорректные данные для постановки лайка.' });
-      } else {
-        res.status(errorCodes.internalServerError).send({ message: err.message });
+        throw new errorsList.BadRequestError('Переданы некорректные данные при удалении карточки.');
       }
-    });
+    })
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -70,16 +72,15 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (card === null) {
-        res.status(errorCodes.notFound).send({ message: 'Карточка не найдена.' });
+        throw new errorsList.NotFoundError('Карточка не найдена.');
       } else {
         res.send({ data: card });
       }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorCodes.badRequest).send({ message: 'Переданы некорректные данные для постановки лайка.' });
-      } else {
-        res.status(errorCodes.internalServerError).send({ message: err.message });
+        throw new errorsList.BadRequestError('Переданы некорректные данные для постановки лайка.');
       }
-    });
+    })
+    .catch(next);
 };
